add react-ga to project. init google analytics and report pageviews on component mount in App.tsx. Reorganize code to minimize amound of functions living in App.tsx
