+++ neuroscout/frontend/package.json
@@ -19,6 +19,7 @@
     "react": "^16.3.2",
     "react-beautiful-dnd": "^10.0.3",
     "react-dom": "^16.3.2",
+    "react-ga": "^2.6.0",
     "react-google-login": "^3.2.1",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
