+++ neuroscout/frontend/src/Bibliography.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { Button, Card, Icon, Dropdown, Skeleton } from 'antd';
 import { config } from './config';
 import { displayError, jwtFetch, alphaSort, timeout } from './utils';
-import './Bibliography.css';
+import './css/Bibliography.css';
 
 const domainRoot = config.server_url;
 const ButtonGroup = Button.Group;
