+++ neuroscout/frontend/src/coretypes.ts
@@ -283,3 +283,12 @@ export interface AuthStoreState {
   gAuth: any;
   avatar: string;
 }
+
+export interface AppState {
+  analyses: AppAnalysis[]; // List of analyses belonging to the user
+  publicAnalyses: AppAnalysis[]; // List of public analyses
+  auth: AuthStoreState;
+  datasets: Dataset[];
+  cloneAnalysis: (number) => void;
+  onDelete:  (analysis: AppAnalysis) => void;
+}
