+++ neuroscout/frontend/src/css/index.css
