+++ neuroscout/frontend/src/index.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import App from './App';
-// import './index.css';
 
 ReactDOM.render(<App />, document.getElementById('root') as HTMLElement);
