+++ neuroscout/frontend/src/utils/analytics.tsx
@@ -1,0 +1,40 @@
+import * as React from 'react';
+import GoogleAnalytics from 'react-ga';
+
+GoogleAnalytics.initialize('UA-143429930-1');
+
+export const withTracker = (WrappedComponent, options = {})  => {
+  const trackPage = page => {
+    GoogleAnalytics.set({
+      page,
+      ...options,
+    });
+    GoogleAnalytics.pageview(page);
+  };
+
+  // eslint-disable-next-line
+  const HOC = class extends React.Component<any, any> {
+    componentDidMount() {
+      // eslint-disable-next-line
+      const page = this.props.location.pathname + this.props.location.search;
+      trackPage(page);
+    }
+
+    componentDidUpdate(prevProps) {
+      const currentPage =
+        prevProps.location.pathname + prevProps.location.search;
+      const nextPage =
+        this.props.location.pathname + this.props.location.search;
+
+      if (currentPage !== nextPage) {
+        trackPage(nextPage);
+      }
+    }
+
+    render() {
+      return <WrappedComponent {...this.props} />;
+    }
+  };
+
+  return HOC;
+};
